feat: datafusion json handler
