+++ datafusion/Cargo.toml
@@ -31,7 +31,8 @@ futures = "0.3"
 itertools = "0.14"
 parking_lot = "0.12"
 tokio = { version = "1", features = ["full"] }
-url = "2.4"
+tracing = "0.1"
+url = "2.5.4"
 
 [dev-dependencies]
 tokio = { version = "1", features = ["full"] }
