+++ datafusion/src/lib.rs
@@ -3,5 +3,6 @@ mod error;
 mod exec;
 mod expressions;
 mod table_provider;
+mod utils;
 
 pub use table_provider::DeltaTableProvider;
